requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
colorama==0.4.6
distro==1.9.0
faiss-cpu==1.12.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.1
numpy==2.3.4
openai==2.7.1
packaging==25.0
pydantic==2.12.4
pydantic_core==2.41.5
PyPDF2==3.0.1
python-dotenv==1.2.1
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
PyPDF2
